+++ src/oliker_prussner_method/core.py
@@ -14,8 +14,8 @@
 import numpy as np
 import numpy.typing as npt
 from numba import jit
-
-import core_src.oliker_prussner_coretriag as optri
+from . import oliker_prussner_coretriag as optri
+# import .oliker_prussner_coretriag as optri
 
 ########################################################################################################
 #   IMPORTANT EQUATION CALCULATIONS
