+++ src/oliker_prussner_method/initialisation.py
@@ -14,7 +14,8 @@ import numpy as np
 import numpy.typing as npt
 from typing import Callable
 from scipy.spatial import Delaunay
-import core_src.oliker_prussner_coretriag as optri
+from . import oliker_prussner_coretriag as optri
+# import core_src.oliker_prussner_coretriag as optri
 
 
 ########################################################################################################
